iBloom - GUIDELINES/requirements.txt
# Your existing requirements
flask>=2.3.0
flask-sqlalchemy>=3.0.0
openai>=1.0.0
python-dotenv>=1.0.0
gunicorn>=21.0.0
psycopg2-binary>=2.9.0
reportlab>=4.0.0
openpyxl>=3.1.0

# New dependencies for clinical AI
sentence-transformers>=2.2.0
chromadb>=0.4.0
PyMuPDF>=1.23.0  # For PDF processing
numpy>=1.24.0
scikit-learn>=1.3.0
langchain>=0.1.0  # For advanced document processing